pyproject.toml
[project]
name = "mealmetrics"
version = "0.1.0"
description = "A report generator for Miran app take-home task"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "django>=5.2.4",
    "django-dotenv>=1.4.2",
    "django-stubs[compatible-mypy]>=5.2.1",
    "djangorestframework>=3.16.0",
    "itsdangerous>=2.2.0",
    "openpyxl>=3.1.5",
    "psycopg2-binary>=2.9.10",
    "redis>=6.2.0",
]

[dependency-groups]
dev = [
    "django-stubs>=5.2.1",
    "isort>=6.0.1",
    "mypy>=1.16.1",
    "mypy-extensions>=1.1.0",
    "pre-commit>=4.2.0",
    "ruff>=0.12.2",
]

[tool.ruff]
line-length = 88
target-version = "py312"
extend-select = ["B", "I", "PL"]
exclude = ["migrations"]

[tool.isort]
profile = "black"
known_first_party = ["reports", "mealsmetrics"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "mealsmetrics.settings"
